requirements.txt
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-Bcrypt==1.0.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gunicorn==25.1.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mercadopago==2.3.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==26.0
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
uritemplate==4.2.0
urllib3==2.6.2
Werkzeug==3.1.5
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.3.3
python-dotenv==1.0.1
